refactor: structure modes by mode
