+++ src/modes/indicator/IndicatorAbout.svelte
