+++ src/modes/indicator/IndicatorCorrelation.svelte
