+++ src/modes/summary/AllIndicatorOverview.svelte
