+++ src/modes/summary/AllIndicatorsText.svelte
