+++ src/modes/summary/GeoTable.svelte
