+++ src/modes/summary/HighlightIndicators.svelte
