+++ src/modes/summary/HistoryLineChart.svelte
