+++ src/modes/summary/HistoryLineTooltip.svelte
