+++ src/modes/summary/IndicatorOverview.svelte
