+++ src/modes/summary/IndicatorStatsLine.svelte
