+++ src/modes/summary/IndicatorTable.svelte
