+++ src/modes/summary/RegionHexMap.svelte
