+++ src/modes/summary/RegionMap.svelte
