+++ src/modes/summary/RegionMapTooltip.svelte
