+++ src/modes/summary/SparkLineTooltip.svelte
