+++ src/modes/survey-results/IndicatorRevisions.svelte
