+++ src/modes/survey-results/SurveyQuestion.svelte
@@ -2,10 +2,10 @@
   import UIKitHint from '../../components/UIKitHint.svelte';
   import fileIcon from '!raw-loader!@fortawesome/fontawesome-free/svgs/solid/file.svg';
   import linkIcon from '!raw-loader!@fortawesome/fontawesome-free/svgs/solid/link.svg';
-  import HistoryLineChart from '../mobile/HistoryLineChart.svelte';
-  import IndicatorOverview from '../mobile/IndicatorOverview.svelte';
+  import HistoryLineChart from '../summary/HistoryLineChart.svelte';
+  import IndicatorOverview from '../summary/IndicatorOverview.svelte';
   import { formatDateYearWeekdayAbbr } from '../../formats';
-  import IndicatorStatsLine from '../mobile/IndicatorStatsLine.svelte';
+  import IndicatorStatsLine from '../summary/IndicatorStatsLine.svelte';
   import IndicatorRevisions from './IndicatorRevisions.svelte';
   import WarningBanner from '../../components/WarningBanner.svelte';
