+++ src/modes/survey-results/SurveyResults.svelte
