+++ src/modes/survey-results/SurveyStats.svelte
